requirements.txt
altgraph==0.17.4
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.11
packaging==25.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
PyQt6==6.10.0
PyQt6-Qt6==6.10.0
PyQt6_sip==13.10.2
PySide6==6.10.0
PySide6_Addons==6.10.0
PySide6_Essentials==6.10.0
python-qbittorrent==0.4.3
requests==2.32.5
setuptools==80.9.0
shiboken6==6.10.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
